+++ RecipeBookWeb/src/components/MultiSelectDropdown/MultiSelect.css
@@ -1,38 +1,32 @@
-.form-container {
-    width: 50% !important;
-    min-width: 450px;
-}
-
-.category-input {
+.multi-select-dropdown {
     border: 1px solid transparent;
     box-shadow: 0px 0px 0px 4px transparent;
     border-radius: 5px;
 }
 
-.category-input:focus-within {
+.multi-select-dropdown:focus-within {
     border: 1px solid #0d6efd80;
     box-shadow: 0px 0px 0px 4px #0d6efd40;
 }
 
-.category-search-label {
+.multi-select-search-bar {
     width: calc(100% - 3.3rem) !important;
 }
 
-.dropdown-button {
+.multi-select-dropdown-button {
     max-width: 3.3rem !important;
 }
 
-.category-search:focus {
+.multi-select-selected-items {
+    background-color: rgb(250, 250, 250);
+}
+
+.multi-select-search-input:focus {
     outline: none;
 }
 
-.category-search {
+.multi-select-search-input {
     min-width: 9ch;
     max-width: 100%;
     border: 4px solid transparent;
-}
-
-
-.selected {
-    background-color: rgb(250, 250, 250);
 }
