+++ RecipeBookWeb/src/pages/AddRecipe/Add.css
@@ -1,0 +1,4 @@
+.form-container {
+    width: 50% !important;
+    min-width: 450px;
+}
